backend/apps/sessions/serializers.py
from rest_framework import serializers
from apps.sessions.models import Session

class SessionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Session
        fields = '__all__'