backend/requirements.txt
# backend/requirements.txt
# Core Django
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Database
psycopg2-binary==2.9.11
dj-database-url==1.3.0

# Security & CORS
django-cors-headers==4.9.0
PyJWT==2.10.1

# Environment & Configuration
django-environ==0.12.0
python-dotenv==1.2.1

# Media & Images
Pillow==12.1.0
qrcode==8.2

# ASGI
asgiref==3.10.0

# Production
gunicorn==24.0.0
whitenoise==6.10.0  # Для обслуживания статических файлов

# Development (можно разделить на requirements-dev.txt)
colorama==0.4.6
packaging==26.0
setuptools==80.10.2
sqlparse==0.5.3
tzdata==2025.2
wheel==0.46.0

# Optional (для будущих функций)
# django-filter==25.3
# django-debug-toolbar==4.5.0
# drf-yasg==1.21.10  # Swagger документация